requirements.txt
google-generativeai>=0.7.0
pypdf>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
tiktoken

